requirements.txt
Django==1.8.6
Markdown==2.6.5
Pillow==2.8.1
argparse==1.3.0
beautifulsoup4==4.4.1
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==1.0.1
django-compressor==1.5
django-crispy-forms==1.5.2
django-debug-toolbar==1.4
django-htmlmin==0.7.0
django-storages-redux==1.2.3
django-toolbelt==0.0.1
djangorestframework==3.3.3
gunicorn==19.3.0
html5lib==0.99999
oauthlib==1.0.3
psycopg2==2.6
pytz==2015.4
requests==2.9.1
requests-oauthlib==0.6.0
six==1.9.0
sqlparse==0.1.18
static3==0.5.1
tweepy==3.5.0
whitenoise==1.0.6
